adding welcome tips

+++ my-app/src/scenes/dashboard/GastosContext.js
@@ -72,14 +72,14 @@ export const GastosProvider = ({ children }) => {
     );
   };
 
-  // Função para agrupar gastos por mês e categoria
+  //Função para agrupar gastos por mês e categoria
   const agruparGastosPorMes = () => {
     // Agrupando gastos por mês
     const gastosPorMes = groupBy(gastos, (gasto) => new Date(gasto.data).getMonth());
 
     // Criando o array de resultado por mês e categorias
     return Object.keys(gastosPorMes).map((mes) => {
-      const grouped = { month: new Date(2023, mes, 1).toLocaleString('default', { month: 'short' }) };
+      const grouped = { month: new Date(2023, mes, 1).toLocaleString('default', { month: 'long' }) };
       
       // Para cada tipo de gasto, somar os valores e adicionar ao objeto
       tiposGasto.forEach((tipo) => {
